Export readDirAndPush function because it was used in multiples files
