+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "",
   "main": "build/index.js",
-  "types": "./typing/index.d.ts",
+  "types": "./typings/index.d.ts",
   "scripts": {
     "build": "tsc",
     "build:typing": "tsc --build tsconfig.typing.json"
