+++ src/handlers/ButtonsHandler.ts
@@ -1,5 +1,5 @@
-import { readdir, stat } from "fs/promises";
 import { join } from "path";
+import { readDirAndPush } from "../util/readDirFiles";
 import type { ShewenyClient } from "../index";
 
 export class ButtonsHandler {
@@ -14,7 +14,7 @@ export class ButtonsHandler {
 
   async registerAll() {
     const baseDir = join(require.main!.path, this.dir);
-    const buttonsPaths: string[] = await this.readDirAndPush(baseDir);
+    const buttonsPaths: string[] = await readDirAndPush(baseDir);
     for (const buttonPath of buttonsPaths) {
       const Button = (await import(buttonPath)).default;
       if (!Button) continue;
@@ -23,23 +23,6 @@ export class ButtonsHandler {
       instance.path = buttonPath;
       this.client.buttons.set(instance.name, instance);
     }
-    return this.client.events;
-  }
-
-  async readDirAndPush(d: string): Promise<Array<string>> {
-    const files: string[] = [];
-    async function read(dir: string) {
-      const result = await readdir(dir);
-      for (const item of result) {
-        const infos = await stat(join(dir, item));
-        if (infos.isDirectory()) await read(join(dir, item));
-        else files.push(join(dir, item));
-      }
-      return;
-    }
-
-    await read(d);
-
-    return files;
+    return this.client.buttons;
   }
 }
