+++ src/handlers/CommandsHandler.ts
@@ -1,6 +1,6 @@
-import { readdir, stat } from "fs/promises";
 import { join } from "path";
 import { SlashHandler } from "../index";
+import { readDirAndPush } from "../util/readDirFiles";
 import type { ICommandHandlerOptions } from "../typescript/interfaces/interfaces";
 import type { ShewenyClient } from "../index";
 
@@ -37,7 +37,7 @@ export class CommandsHandler {
    */
   async loadAll() {
     const baseDir = join(require.main!.path, this.dir);
-    const cmds: string[] = await this._readDirAndPush(baseDir);
+    const cmds: string[] = await readDirAndPush(baseDir);
     for (const cmdPath of cmds) {
       const commandImport = await import(cmdPath);
       const key = Object.keys(commandImport)[0];
@@ -54,26 +54,5 @@ export class CommandsHandler {
       this.slashCommands = new SlashHandler(this.client);
     }
     return this.client.commands;
-  }
-  /**
-   * Read dir and return array with all paths of files
-   * @param {string} directory - The directory to read
-   * @returns {Array<string>}
-   */
-  async _readDirAndPush(d: string): Promise<Array<string>> {
-    const files: string[] = [];
-    async function read(dir: string) {
-      const result = await readdir(dir);
-      for (const item of result) {
-        const infos = await stat(join(dir, item));
-        if (infos.isDirectory()) await read(join(dir, item));
-        else files.push(join(dir, item));
-      }
-      return;
-    }
-
-    await read(d);
-
-    return files;
   }
 }
