+++ src/handlers/EventsHandler.ts
@@ -1,5 +1,5 @@
-import { readdir, stat } from "fs/promises";
 import { join } from "path";
+import { readDirAndPush } from "../util/readDirFiles";
 import type { ShewenyClient } from "../index";
 
 /**
@@ -26,7 +26,7 @@ export class EventsHandler {
    */
   async registerAll() {
     const baseDir = join(require.main!.path, this.dir);
-    const evtsPaths: string[] = await this._readDirAndPush(baseDir);
+    const evtsPaths: string[] = await readDirAndPush(baseDir);
     for (const evtPath of evtsPaths) {
       const Event = (await import(evtPath)).default;
       if (!Event) continue;
@@ -46,26 +46,5 @@ export class EventsHandler {
     for (const [name, evt] of this.client.events) {
       this.client.on(name, (...args: any[]) => evt.execute(args));
     }
-  }
-  /**
-   * Read dir and return array with all paths of files
-   * @param {string} directory - The directory to read
-   * @returns {Array<string>}
-   */
-  async _readDirAndPush(d: string): Promise<Array<string>> {
-    const files: string[] = [];
-    async function read(dir: string) {
-      const result = await readdir(dir);
-      for (const item of result) {
-        const infos = await stat(join(dir, item));
-        if (infos.isDirectory()) await read(join(dir, item));
-        else files.push(join(dir, item));
-      }
-      return;
-    }
-
-    await read(d);
-
-    return files;
   }
 }
