+++ test/src/commands/ping.ts
@@ -1,15 +1,15 @@
-import { Command, ShewenyClient } from '../../../';
-import type { Message } from 'discord.js'
+import { Command, ShewenyClient } from "../../../";
+import type { Message } from "discord.js";
 
 export class PingCommand extends Command {
-	constructor(client: ShewenyClient) {
-		super(client, 'ping', {
-			description: 'Ping the bot',
-			category: 'Misc',
-			DMOnly: true,
-		})
-	}
-	execute(message: Message) {
-		message.reply('Pong !')
-	}
+  constructor(client: ShewenyClient) {
+    super(client, "ping", {
+      description: "Ping the bot",
+      category: "Misc",
+      guildOnly: true,
+    });
+  }
+  execute(message: Message) {
+    message.reply("Pong !");
+  }
 }
